Add authentication docs (#2683)
Now that we've added support for UMI and FIC, this PR aims to clean up
some of our docs around auth.

Note we still don't have support for it in C# yet (but soon!)

1. Our M365Agents Toolkit and Manual Configuration guides were floating
around. I moved it to a Basic Configuration Section
2. I added an App Authentication section in Teams Integration section to
show how to set up various types of auth
3. I added an App Authentication section in Essentials to show how to
enable each of the types of app authentication
4. I moved Teams Integration above the Language Guide section because
it's a precursor to doing anything in code.
5. For Manual Configuration, I used Tabs to showcase both Azure Portal
and Azure CLI

<img width="324" height="1151" alt="Screenshot 2025-11-13 at 3 33 29 PM"
src="https://github.com/user-attachments/assets/576ebee8-98ea-444e-8252-b96095ba9d09"
/>